Check for .env file & Change .handlebars to .hbs

+++ index.js
@@ -2,10 +2,16 @@
 var config = require('dotenv').config();
 var fs = require('fs');
 var config_example = "";
-for (var attributename in config.parsed) {
-    config_example += attributename + "=\r\n";
+if (fs.existsSync("./.env")) {
+    for (var attributename in config.parsed) {
+        config_example += attributename + "=\r\n";
+    }
+    fs.writeFileSync('./.env.example', config_example);
+} else {
+    //fs.copyFileSync("./.env.example", ".env");
+    console.log("Update .env Files first!");
+    process.exit(1);
 }
-fs.writeFileSync('./.env.example', config_example);
 /* Example File Finish */
 
 var db = require("./db.js");
@@ -29,10 +35,9 @@ var express = require('express');
 var exphbs = require('express-handlebars');
 var i18n = require("i18n");
 
-
 var app = express();
 
-app.engine('handlebars', exphbs({ defaultLayout: 'main' }));
+app.engine('handlebars', exphbs({ extname: '.hbs', defaultLayout: 'main' }));
 app.set('view engine', 'handlebars');
 
 i18n.configure({
